web/requirements.txt
djangorestframework
django-filter
dj-rest-auth